MP_Speech_Scraper/requirements.txt
requests==2.31.0
beautifulsoup4==4.12.2
python-dateutil==2.8.2
tqdm==4.66.1
lxml==5.1.0 